replace fluent img
