+++ packages/website-astro/src/layouts/showcase.astro
@@ -5,6 +5,7 @@ import BaseLayout from "./base-layout.astro";
 <style>
   .showcase {
     background-color: var(--colorNeutralBackground3);
+    --ec-brdWd: 0px;
   }
 </style>
